requirements.txt
fastapi
uvicorn
pydantic
stripe
PyYAML
requests
httpx
transformers
jsonschema
SQLAlchemy
psycopg2-binary
chromadb
weaviate-client
jinja2
email-validator
passlib[bcrypt]
python-jose[cryptography]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
boto3
python-dotenv
